employee_microservice/src/core/config.py
import os
from dotenv import load_dotenv
from fastapi.security import OAuth2PasswordBearer

load_dotenv()

SECRET_KEY = os.getenv("SECRET_KEY")
ALGORITHM = "HS256"

oauth2 = OAuth2PasswordBearer(tokenUrl="/token")