employee_microservice/pyproject.toml
[project]
name = "employee-microservice"
version = "0.1.0"
description = ""
authors = [
    {name = "OliverRoat",email = "oliver.roat@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "aio-pika (>=9.5.5,<10.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.3,<3.0.0)",
    "pika (>=1.3.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
